// api/models/user-profile-with-relation-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Social Media API
 * The Social Media API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { IsBlockedDTO } from './is-blocked-dto';
import { IsFollowedDTO } from './is-followed-dto';
import { TranslateResultDto } from './translate-result-dto';
 /**
 * 
 *
 * @export
 * @interface UserProfileWithRelationDTO
 */
export interface UserProfileWithRelationDTO {

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    id: string;

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    name: string;

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    username: string;

    /**
     * @type {TranslateResultDto}
     * @memberof UserProfileWithRelationDTO
     */
    about: TranslateResultDto;

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    profilePicture: string;

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    profilePictureBlurhash?: string;

    /**
     * @type {number}
     * @memberof UserProfileWithRelationDTO
     */
    followerCount: number;

    /**
     * @type {number}
     * @memberof UserProfileWithRelationDTO
     */
    followingCount: number;

    /**
     * @type {number}
     * @memberof UserProfileWithRelationDTO
     */
    postCount: number;

    /**
     * @type {string}
     * @memberof UserProfileWithRelationDTO
     */
    language: UserProfileWithRelationDTOLanguageEnum;

    /**
     * @type {IsFollowedDTO}
     * @memberof UserProfileWithRelationDTO
     */
    followStatus: IsFollowedDTO;

    /**
     * @type {IsBlockedDTO}
     * @memberof UserProfileWithRelationDTO
     */
    blockStatus: IsBlockedDTO;
}

/**
 * @export
 * @enum {string}
 */
export enum UserProfileWithRelationDTOLanguageEnum {
    En = 'en',
    Es = 'es',
    Fr = 'fr',
    De = 'de',
    It = 'it',
    PtPT = 'pt-PT',
    PtBR = 'pt-BR',
    Ru = 'ru',
    Ja = 'ja',
    ZhCN = 'zh-CN',
    ZhTW = 'zh-TW',
    Ko = 'ko',
    Ar = 'ar',
    Hi = 'hi',
    Tr = 'tr',
    Nl = 'nl',
    Pl = 'pl',
    Uk = 'uk',
    El = 'el',
    Id = 'id',
    Sv = 'sv',
    Cs = 'cs',
    Fi = 'fi',
    Ro = 'ro',
    No = 'no',
    Bg = 'bg',
    Sr = 'sr',
    Hr = 'hr',
    Bs = 'bs',
    Sl = 'sl'
}